src/fixforheroku.py
from dotenv import load_dotenv
import os
import re

load_dotenv()

uri = os.getenv("DATABASE_URL")
if uri.startswith("postgres://"):
    uri = uri.replace("postgres://", "postgresql://", 1)
# rest of connection code using the connection string `uri`