pyproject.toml
[tool.poetry]
name = "ohtuminiproject"
version = "0.1.0"
description = ""
authors = [
    "Pentti Lähdeaho <pentti.lahdeaho@gmail.com>",
    "Your Name <you@example.com>"
    ]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
gunicorn = "^20.1.0"
robotframework = "^4.1.2"
robotframework-seleniumlibrary = "^5.1.3"
requests = "^2.26.0"
Flask-SQLAlchemy = "^2.5.1"
psycopg2 = "^2.9.2"
python-dotenv = "^0.19.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
coverage = "^6.2"
pylint = "^2.12.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
